+++ www/painel/Models/Options.php
@@ -6,10 +6,17 @@ use \Core\Model;
 
 class Options extends Model {
 
-	public function getAll():array {
+	public function getAll($checkHasProduct = false):array {
     $array = array();
 
-    $sql = "SELECT * FROM options";
+    if($checkHasProduct){
+      $sql = "SELECT *, 
+                (select count(*) from products_options where products_options.id = options.id) as product_count 
+                  FROM options";
+    }else{
+      $sql = "SELECT * FROM options";
+    }
+
     $sql = $this->db->query($sql);
 
     if($sql->rowCount() > 0){
@@ -56,11 +63,18 @@ class Options extends Model {
 
   public function delete($id){
 
-    $sql = "DELETE FROM options WHERE id=:id";
+    $sql = "SELECT count(*) as c FROM products_options WHERE id_option = :id";
     $sql = $this->db->prepare($sql);
     $sql->bindValue(":id", $id);
     $sql->execute();
+    $data = $sql->fetch();
 
+    if($data['c'] == '0'){
+      $sql = "DELETE FROM options WHERE id=:id";
+      $sql = $this->db->prepare($sql);
+      $sql->bindValue(":id", $id);
+      $sql->execute();
+    }
   }
 
 }
